Fix feature hint for DLC checkboxes
